LINT: fix "Assignment of ssize_t to int" warnings
git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@1696 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ exec/crypto.c
@@ -1250,11 +1250,11 @@ static unsigned long rng_nix(unsigned char *buf, unsigned long len,
                              void (*callback)(void))
 {
 	int fd;
-	int rb;
+	unsigned long rb;
 
 	fd = open ("/dev/urandom", O_RDONLY);
 	
-	rb = read (fd, buf, len);
+	rb = (unsigned long)read (fd, buf, len);
 
 	close (fd);
